refactor(menu-bar): split MenuBarView into focused collaborators (C1c)
MenuBarView.swift was 958 lines and owned four unrelated concerns: the
popover shell, the overview panel, the provider status card, and the
popover's sizing math. Audit item C1c.

The sizing math was the real defect. The view carried four private
geometry constants and three private clamp properties, and
notifyContentSize re-derived the same clamp chain by hand — so the
height the popover rendered at and the height it reported to the
menu-bar window controller were two independent copies of one rule,
free to drift. Both now call MenuBarPopoverGeometry's functions, and
the agreement is assertable without hosting a view.

Extracted as collaborators with explicit inputs rather than cross-file
extensions: Swift `private` is file-scoped, so extensions would have
forced weakening every private member.

- MenuBarPopoverGeometry: sizing statics + maximumHeight/scrollHeight/
  popoverHeight/contentSize
- ProviderStatusDotsSummary + PopoverHeaderStatusDots
- PopoverCardID / PopoverCardFramesPreferenceKey / reportPopoverCardFrame
- ProviderCardPresentation: status color/text, detail-navigation and
  reset-credit predicates, fable accessibility label
- ProviderCardResetCreditOutcome: alert mapping that keeps a failed
  post-redemption refresh from reading like a failed redemption
- ProviderCardFableActivityRow: scopes its 60s TimelineView ticker so a
  minute boundary no longer invalidates the whole card
- PopoverOverviewPanel, ProviderStatusCard

MenuBarView 958 -> 285 lines. No call-site signatures change;
ProviderStatusCard.init keeps onSelect last so trailing-closure sites
bind unchanged. Adds 28 test cases across 5 suites.

Deliberately not merged: the three provider-status summary strings
(dots / detail content / dashboard section) differ in states, guards,
and copy; the reasoning is recorded in ProviderStatusDotsSummary.

Verified: SwiftLint --strict clean; mechanical declaration diff shows
all 12 relocated declarations accounted for. Tests/build not run
locally; PR CI is the gate.
