+++ MeterBar/Views/ProviderStatusCard.swift
@@ -1,0 +1,315 @@
+import Foundation
+import MeterBarShared
+import SwiftUI
+
+// Non-private so the shared card shell can be rendered in both usage states by
+// `LiquidGlassP1RegressionTests`.
+struct ProviderStatusCard: View {
+    let snapshot: ProviderSnapshot
+    var onSelect: (() -> Void)?
+    var onHoverChange: ((Bool) -> Void)?
+
+    @ObservedObject private var menuBarDisplayPreferences = MenuBarDisplayPreferencesStore.shared
+    @Environment(\.accessibilityReduceMotion)
+    private var reduceMotion
+    @StateObject private var codexService = CodexCliLocalService.shared
+    @StateObject private var codexAccounts = CodexAccountStore.shared
+    @StateObject private var dataManager = UsageDataManager.shared
+    @State private var isCodexAuthenticated = false
+    @State private var isConsumingResetCredit = false
+    @State private var didConsumeResetCredit = false
+    @State private var showingResetCreditConfirmation = false
+    @State private var resetCreditAlertTitle = ProviderCardResetCreditOutcome.failureTitle
+    @State private var resetCreditAlertMessage: String?
+
+    // `onSelect` stays last so trailing-closure call sites keep binding to it.
+    init(
+        snapshot: ProviderSnapshot,
+        onHoverChange: ((Bool) -> Void)? = nil,
+        onSelect: (() -> Void)? = nil
+    ) {
+        self.snapshot = snapshot
+        self.onHoverChange = onHoverChange
+        self.onSelect = onSelect
+    }
+
+    private var statusColor: Color {
+        ProviderCardPresentation.statusColor(for: snapshot)
+    }
+
+    private var statusText: String {
+        ProviderCardPresentation.statusText(for: snapshot)
+    }
+
+    /// Cards without usage data and exhausted cards are terminal summaries. A
+    /// login/waiting card has no quota detail to reveal, while an exhausted card
+    /// already shows its only actionable reset information inline.
+    var allowsDetailNavigation: Bool {
+        ProviderCardPresentation.allowsDetailNavigation(hasSelectionHandler: onSelect != nil, snapshot: snapshot)
+    }
+
+    var body: some View {
+        selectableCard
+            .providerCardContextMenu(ProviderCardCommands.standard(snapshot: snapshot))
+            .task(id: snapshot.updatedAt) {
+                await refreshCodexAuthenticationState()
+            }
+            .confirmationDialog(
+                "Use a Codex reset credit?",
+                isPresented: $showingResetCreditConfirmation,
+                titleVisibility: .visible
+            ) {
+                Button("Use Reset Credit") { consumeResetCredit() }
+                Button("Cancel", role: .cancel) {}
+            } message: {
+                Text(
+                    "This spends one of your finite Codex reset credits and resets the active blocked usage window. " +
+                        "The action cannot be undone."
+                )
+            }
+            .alert(
+                resetCreditAlertTitle,
+                isPresented: Binding(
+                    get: { resetCreditAlertMessage != nil },
+                    set: { if !$0 { resetCreditAlertMessage = nil } }
+                )
+            ) {
+                Button("OK") { resetCreditAlertMessage = nil }
+            } message: {
+                Text(resetCreditAlertMessage ?? "")
+            }
+    }
+
+    @ViewBuilder private var selectableCard: some View {
+        if let onSelect, allowsDetailNavigation {
+            Button(action: onSelect) {
+                cardContent
+            }
+            .buttonStyle(ProviderCardButtonStyle())
+            .accessibilityHint("Open \(snapshot.title) provider details")
+            .onHover { onHoverChange?($0) }
+        } else {
+            cardContent
+        }
+    }
+
+    /// Chevron shown only when the card opens a detail panel, so "clickable" is
+    /// visible instead of relying on an accessibilityHint alone.
+    @ViewBuilder private var disclosureChevron: some View {
+        if allowsDetailNavigation {
+            CardDisclosureChevron()
+        }
+    }
+
+    private var cardContent: some View {
+        DashboardTile(padding: 11) {
+            cardBody
+        }
+        .contentShape(RoundedRectangle(cornerRadius: MeterBarTheme.Radius.card, style: .continuous))
+    }
+
+    @ViewBuilder private var cardBody: some View {
+        if !snapshot.hasMetrics {
+            offlineRow
+        } else if snapshot.hasExhaustedWeeklyLimit {
+            weeklyExhaustedRow
+        } else {
+            expandedCardBody
+        }
+    }
+
+    /// Logged-out/no-data providers are terminal status rows. Recovery belongs in
+    /// Settings; the popover only needs to say that this profile is offline.
+    private var offlineRow: some View {
+        HStack(spacing: 7) {
+            ProviderLogoView(kind: snapshot.logoKind, size: 17, foregroundColor: snapshot.accentColor)
+            Text(snapshot.title)
+                .font(.subheadline)
+                .fontWeight(.semibold)
+                .lineLimit(1)
+            Spacer(minLength: 8)
+            Text("Offline")
+                .font(.caption2)
+                .fontWeight(.semibold)
+                .foregroundStyle(.secondary)
+        }
+    }
+
+    /// A weekly block makes every shorter/model-specific gauge non-actionable.
+    /// Keep the shared card surface, but collapse its content to the provider and
+    /// the one reset that can restore service.
+    private var weeklyExhaustedRow: some View {
+        TimelineView(.periodic(from: ResetCountdownSchedule.anchor, by: ResetCountdownSchedule.interval)) { timeline in
+            let blockingWindow = BlockingLimitResetCounter.selectBlockingWindow(
+                snapshot.resetWindows,
+                now: timeline.date
+            )
+            let title = BlockingLimitResetCounter.titleText(for: blockingWindow, in: snapshot.resetWindows)
+            let counter = BlockingLimitResetCounter.counterText(
+                for: blockingWindow,
+                now: timeline.date,
+                format: menuBarDisplayPreferences.resetTimeFormat
+            )
+
+            HStack(spacing: 7) {
+                ProviderLogoView(kind: snapshot.logoKind, size: 17, foregroundColor: snapshot.accentColor)
+                Text(snapshot.title)
+                    .font(.subheadline)
+                    .fontWeight(.semibold)
+                    .lineLimit(1)
+
+                Spacer(minLength: 8)
+
+                Text(statusText)
+                    .font(.caption2)
+                    .fontWeight(.semibold)
+                    .foregroundColor(statusColor)
+
+                Label("\(title) \(counter)", systemImage: "hourglass")
+                    .font(.caption)
+                    .fontWeight(.semibold)
+                    .foregroundColor(snapshot.accentColor)
+                    .monospacedDigit()
+                    .lineLimit(1)
+                    .minimumScaleFactor(0.72)
+                    .numericRefreshTransition(value: counter, reduceMotion: reduceMotion)
+
+                if showsResetCreditAction {
+                    Button {
+                        showingResetCreditConfirmation = true
+                    } label: {
+                        Image(systemName: "arrow.clockwise.circle.fill")
+                    }
+                    .buttonStyle(.borderless)
+                    .help("Use a Codex reset credit")
+                    .disabled(isConsumingResetCredit)
+                }
+            }
+        }
+    }
+
+    private var expandedCardBody: some View {
+        VStack(alignment: .leading, spacing: 10) {
+            HStack(spacing: 7) {
+                ProviderLogoView(kind: snapshot.logoKind, size: 17, foregroundColor: snapshot.accentColor)
+                VStack(alignment: .leading, spacing: 1) {
+                    Text(snapshot.title)
+                        .font(.subheadline)
+                        .fontWeight(.semibold)
+                        .lineLimit(1)
+                    Text(snapshot.updatedText)
+                        .font(.caption2)
+                        .foregroundColor(.secondary)
+                        .lineLimit(1)
+                }
+                Spacer(minLength: 0)
+                Text(statusText)
+                    .font(.caption2)
+                    .fontWeight(.semibold)
+                    .foregroundColor(statusColor)
+
+                disclosureChevron
+            }
+
+            if snapshot.hasExhaustedLimit {
+                BlockingLimitResetCounter(
+                    windows: snapshot.resetWindows,
+                    accentColor: snapshot.accentColor,
+                    format: menuBarDisplayPreferences.resetTimeFormat
+                )
+            } else {
+                VStack(alignment: .leading, spacing: 9) {
+                    ForEach(snapshot.limits) { limit in
+                        LimitRow(limit: limit, accentColor: snapshot.accentColor, density: .compact)
+                    }
+                }
+            }
+
+            fableActivitySection
+
+            let badges = ProviderStatusBadges(snapshot: snapshot, style: .compact)
+            if badges.hasContent {
+                badges
+            }
+
+            if showsResetCreditAction {
+                Divider()
+                resetCreditButton
+            }
+        }
+    }
+
+    @ViewBuilder private var fableActivitySection: some View {
+        if let fableActivity = snapshot.fableActivity {
+            Divider()
+            ProviderCardFableActivityRow(activity: fableActivity)
+        }
+    }
+
+    private var resetCreditButton: some View {
+        Button {
+            showingResetCreditConfirmation = true
+        } label: {
+            HStack(spacing: 6) {
+                if isConsumingResetCredit {
+                    ProgressView()
+                        .controlSize(.small)
+                } else {
+                    Image(systemName: "arrow.clockwise.circle.fill")
+                }
+                Text(isConsumingResetCredit ? "Using reset credit…" : "Use reset credit")
+            }
+            .frame(maxWidth: .infinity)
+        }
+        .buttonStyle(.borderedProminent)
+        .controlSize(.small)
+        .disabled(isConsumingResetCredit)
+    }
+
+    private var showsResetCreditAction: Bool {
+        ProviderCardPresentation.showsResetCreditAction(
+            snapshot: snapshot,
+            didConsumeResetCredit: didConsumeResetCredit,
+            isAuthenticated: isCodexAuthenticated
+        )
+    }
+
+    private var codexAccount: CodexAccount? {
+        guard snapshot.service == .codexCli, let accountID = snapshot.accountID else { return nil }
+        return codexAccounts.accounts.first { $0.id == accountID }
+    }
+
+    private func refreshCodexAuthenticationState() async {
+        guard let codexAccount else {
+            isCodexAuthenticated = false
+            return
+        }
+        isCodexAuthenticated = await codexService.canAccess(account: codexAccount)
+    }
+
+    private func consumeResetCredit() {
+        guard let codexAccount, !isConsumingResetCredit else { return }
+        isConsumingResetCredit = true
+
+        Task {
+            do {
+                let result = try await codexService.consumeResetCredit(account: codexAccount)
+                didConsumeResetCredit = true
+                if let refreshedMetrics = result.refreshedMetrics {
+                    dataManager.applyCodexResetCreditRefresh(refreshedMetrics, accountID: codexAccount.id)
+                }
+                if let alert = ProviderCardResetCreditOutcome.alert(for: result) {
+                    present(alert)
+                }
+            } catch {
+                present(ProviderCardResetCreditOutcome.alert(for: error))
+            }
+            isConsumingResetCredit = false
+        }
+    }
+
+    private func present(_ alert: ProviderCardResetCreditOutcome.Alert) {
+        resetCreditAlertTitle = alert.title
+        resetCreditAlertMessage = alert.message
+    }
+}
